chore(deps): bump jni from 0.21.1 to 0.22.4 in /native

+++ native/core/src/lib.rs
@@ -150,7 +150,7 @@ pub extern "system" fn Java_org_apache_comet_NativeBase_isFeatureEnabled(
             _ => false, // Unknown features return false
         };
 
-        Ok(enabled as u8)
+        Ok(enabled)
     })
 }
